reorganized source for wwivutil to make it easier to keep adding more commands into it.
