+++ wwivutil/CMakeLists.txt
@@ -7,11 +7,11 @@ set(WWIVUTIL_MAIN wwivutil.cpp)
 
 set(COMMAND_SOURCES
   command.cpp
-  dump_callout.cpp
-  dump_contact.cpp
-  dump_packet.cpp
-  messages.cpp
-  net.cpp
+  messages/messages.cpp
+  net/dump_callout.cpp
+  net/dump_contact.cpp
+  net/dump_packet.cpp
+  net/net.cpp
   )
 
 add_executable(wwivutil ${WWIVUTIL_MAIN} ${COMMAND_SOURCES})
