+++ wwivutil/net/dump_packet.cpp
@@ -15,7 +15,7 @@
 /*    either  express  or implied.  See  the  License for  the specific   */
 /*    language governing permissions and limitations under the License.   */
 /**************************************************************************/
-#include "wwivutil/dump_packet.h"
+#include "wwivutil/net/dump_packet.h"
 
 #include <iostream>
 #include <string>
